cpu renderer and math

+++ RadianceFields/Utility/BMPSaver.cpp
@@ -1,6 +1,6 @@
 #include "BMPSaver.h"
 
-void Utility::BMPSaver::Save(std::filesystem::path path, const std::vector<uint8_t>& imageData)
+void Utility::BMPSaver::Save(std::filesystem::path path, const std::vector<float>& imageData)
 {
 
 }
